+++ services/account/scripts/sync-waf-exceptions.mjs
@@ -1,11 +1,17 @@
-import fs from 'node:fs';
 import path from 'node:path';
 import process from 'node:process';
 import { routeDefinitions } from './wrangler-config.mjs';
+import {
+  cloudflare,
+  groupByZoneName,
+  normalizeEnvironment,
+  readAccountConfig,
+  resolveZoneIdForRouteZone
+} from './cloudflare-sync.mjs';
 
 const managedFirewallPhase = 'http_request_firewall_managed';
 const serviceRoot = path.resolve(import.meta.dirname, '..');
-const config = JSON.parse(fs.readFileSync(path.join(serviceRoot, 'config/environments.json'), 'utf8'));
+const config = readAccountConfig(serviceRoot);
 const environment = normalizeEnvironment(process.argv[2] ?? 'dev');
 const regionsArg = process.argv.find((arg) => arg.startsWith('--regions='))?.split('=')[1];
 const regions = regionsArg ? regionsArg.split(',').map((entry) => entry.trim()).filter(Boolean) : Object.keys(config.regions);
@@ -25,7 +31,7 @@ const token = process.env.CLOUDFLARE_WAF_TOKEN
 
 if (!token) {
   if (dryRun) {
-    for (const [zoneName, exceptions] of groupExceptionsByZone(wafExceptionDefinitions(config, environment, regions, productDomains))) {
+    for (const [zoneName, exceptions] of groupByZoneName(wafExceptionDefinitions(config, environment, regions, productDomains))) {
       console.log(`dry-run ${zoneName}`);
       for (const exception of exceptions) {
         console.log(JSON.stringify(exception.rule, null, 2));
@@ -38,10 +44,10 @@ if (!token) {
   process.exit(1);
 }
 
-const exceptionsByZone = groupExceptionsByZone(wafExceptionDefinitions(config, environment, regions, productDomains));
+const exceptionsByZone = groupByZoneName(wafExceptionDefinitions(config, environment, regions, productDomains));
 
 for (const [zoneName, exceptions] of exceptionsByZone) {
-  const zoneId = await resolveZoneIdForRouteZone(zoneName, token);
+  const zoneId = await resolveZoneIdForRouteZone(config, zoneName, token);
   const entrypoint = await readManagedRulesEntrypoint(zoneId, token);
   if (!entrypoint) {
     throw new Error(`Cloudflare managed WAF entrypoint not found for ${zoneName}. Enable the managed ruleset first, then sync exceptions.`);
@@ -121,16 +127,6 @@ function parseProductDomains(config, value) {
   return domains;
 }
 
-function groupExceptionsByZone(exceptions) {
-  const grouped = new Map();
-  for (const exception of exceptions) {
-    const entries = grouped.get(exception.zoneName) ?? [];
-    entries.push(exception);
-    grouped.set(exception.zoneName, entries);
-  }
-  return grouped;
-}
-
 async function readManagedRulesEntrypoint(zoneId, token) {
   try {
     return await cloudflare(`/zones/${zoneId}/rulesets/phases/${managedFirewallPhase}/entrypoint`, token);
@@ -183,55 +179,10 @@ function removeUndefinedValues(value) {
   return Object.fromEntries(Object.entries(value).filter((entry) => entry[1] !== undefined));
 }
 
-async function resolveZoneIdForRouteZone(zoneName, token) {
-  const specificEnvName = `CLOUDFLARE_ZONE_ID_${zoneName.toUpperCase().replace(/[^A-Z0-9]+/g, '_')}`;
-  const specificZoneId = process.env[specificEnvName];
-  if (specificZoneId) return specificZoneId;
-
-  if (zoneName === config.accountDomain && process.env.CLOUDFLARE_ZONE_ID) {
-    return process.env.CLOUDFLARE_ZONE_ID;
-  }
-
-  return resolveZoneId(zoneName, token);
-}
-
-async function resolveZoneId(zoneName, token) {
-  const zones = await cloudflare(`/zones?name=${encodeURIComponent(zoneName)}`, token);
-  const zone = zones.find((entry) => entry.name === zoneName);
-  if (!zone) {
-    throw new Error(`Cloudflare zone not found: ${zoneName}`);
-  }
-  return zone.id;
-}
-
-async function cloudflare(pathname, token, options = {}) {
-  const response = await fetch(`https://api.cloudflare.com/client/v4${pathname}`, {
-    method: options.method ?? 'GET',
-    headers: {
-      authorization: `Bearer ${token}`,
-      'content-type': 'application/json'
-    },
-    body: options.body ? JSON.stringify(options.body) : undefined
-  });
-  const payload = await response.json();
-  if (!payload.success) {
-    const error = new Error(`Cloudflare API failed: ${JSON.stringify(payload.errors)}`);
-    error.cloudflareStatus = response.status;
-    throw error;
-  }
-  return payload.result;
-}
-
 function cloudflareStringSet(values) {
   return `{${values.map(quoteCloudflareString).join(' ')}}`;
 }
 
 function quoteCloudflareString(value) {
   return `"${String(value).replaceAll('\\', '\\\\').replaceAll('"', '\\"')}"`;
-}
-
-function normalizeEnvironment(value) {
-  const normalized = value.trim().toLowerCase();
-  if (['dev', 'pre', 'live'].includes(normalized)) return normalized;
-  return null;
 }
